moved route class to a dedicated ns
